getclnthandle(): Fix assert() condition and destroy client handle.
Currently in getclnthandle() even if 'clnt_tli_ncreate' fails the 'client'
pointer is valid and has the error code saved in 'client->cl_error.re_status'.
But with this later on ganesha aborts for assert(client == NULL).
To fix this we now check if 'client' is not NULL then
assert(CLNT_FAILURE(client)) and then destroy the client as after this
clnt_raw_ncreate() is called to create a client.

Signed-off-by: Madhu Thorat <madhu.punjabi@in.ibm.com>